+++ ui/components/checkbox-toggle/RELEASENOTES.md
@@ -4,6 +4,16 @@
 
 <!-- ## [Unreleased] -->
 
+## 2.28.0
+
+### Fixed
+
+- Fixed checkbox clickable area to only apply within container.
+
+### Added
+
+- Added `role="switch"` to input element to ensure accessibility.
+
 ## 2.27.1
 
 ### Fixed